Move reindex rake task to migration folder
Move reindex rake task to migration folder, add reviewer feedback
and improve reindexing workflow.
